probably done with cleaning training code
