+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__
 wandb/**
 *.pth
+torch_compile_debug/**
